+++ src/command/get.rs
@@ -7,20 +7,28 @@ Slate: Get a value by name.
 
 Usage:
   slate get <key>
-  slate get [options]
+  slate get -n <key>
+  slate get -h
 
 Options:
-  -h --help  Show this help.
+  -h --help    Show this help.
+  -n --no-eol  Do not print the trailing newline character. Show the value as it was saved.
 
 Examples:
 
-  slate get foo
-  #=> bar
+  $ slate get foo
+  bar
+  $ _
+
+  slate get -n foo
+  bar%
+  $ _ # no EOL
 ";
 
 #[derive(Debug, RustcDecodable)]
 struct Args {
     arg_key: String,
+    flag_no_eol: bool,
 }
 
 pub fn run(argv: &Vec<String>) -> Result<Option<Message>, Message> {
@@ -28,7 +36,15 @@ pub fn run(argv: &Vec<String>) -> Result<Option<Message>, Message> {
     let slate: Slate = Default::default();
 
     match slate.get(&args.arg_key) {
+        Ok(value) => {
+            let message: Message = if args.flag_no_eol {
+                Message::Raw(value)
+            } else {
+                Message::Info(value)
+            };
+
+            Ok(Some(message))
+        },
         Err(e) => Err(Message::Info(e.to_owned())),
-        Ok(value) => Ok(Some(Message::Raw(value))),
     }
 }